리팩토링: 레시피 & 커뮤니티 ListPage 공통화
